feat: simple counter example, orig. counter moved to counter-advanced
