+++ examples/counter_advanced/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "counter_advanced"
+version = "0.1.0"
+authors = ["David O'Connor <david.alan.oconnor@gmail.com>"]
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+seed = {path = "../../"}
+wasm-bindgen = "0.2.55"
