+++ examples/counter_advanced/Makefile.toml
@@ -1,0 +1,22 @@
+extend = "../../Makefile.toml"
+
+# ---- BUILD ----
+
+[tasks.build]
+alias = "default_build"
+
+[tasks.build_release]
+alias = "default_build_release"
+
+# ---- START ----
+
+[tasks.start]
+alias = "default_start"
+
+[tasks.start_release]
+alias = "default_start_release"
+
+# ---- LINT ----
+
+[tasks.clippy]
+alias = "default_clippy"
