+++ examples/counter_advanced/README.md
@@ -1,0 +1,11 @@
+## Counter-advanced example
+
+Intended as a more sophisticated demo of basic functionality.
+
+---
+
+```bash
+cargo make start
+```
+
+Open [127.0.0.1:8000](http://127.0.0.1:8000) in your browser.
