+++ examples/counter_advanced/index.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, shrink-to-fit=no"
+    />
+
+    <meta name="description" content="" />
+
+    <link rel="icon" type="image/png" href="/public/favicon.png" />
+
+    <!--<link rel="stylesheet" type="text/css" href="/style.css">-->
+
+    <title>Counter-advanced example</title>
+
+    <!-- Because of Edge, see https://github.com/samthor/fast-text-encoding -->
+    <script type="text/javascript" src="/public/text-polyfill.min.js"></script>
+  </head>
+  <body>
+    <section id="app"></section>
+    <script type="module">
+      // https://rustwasm.github.io/docs/wasm-bindgen/examples/without-a-bundler.html
+      import init from '/pkg/package.js';
+      init('/pkg/package_bg.wasm');
+    </script>
+  </body>
+</html>
